refactor: 메뉴 BC에 따른 패키지 구조 리팩터링
