+++ src/main/java/kitchenpos/menus/application/MenuGroupService.java
@@ -1,7 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.menus.application;
 
-import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRepository;
+import kitchenpos.menus.domain.model.MenuGroup;
+import kitchenpos.menus.domain.model.MenuGroupRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
