+++ src/main/java/kitchenpos/menus/domain/model/MenuGroup.java
@@ -1,9 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.menus.domain.model;
 
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
+import jakarta.persistence.*;
 
 import java.util.UUID;
