+++ src/main/java/kitchenpos/menus/domain/model/MenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menus.domain.model;
 
 import java.util.List;
 import java.util.Optional;
