+++ src/main/java/kitchenpos/menus/infrastructure/JpaMenuGroupRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.menus.infrastructure;
 
+import kitchenpos.menus.domain.model.MenuGroup;
+import kitchenpos.menus.domain.model.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
