+++ src/main/java/kitchenpos/menus/infrastructure/JpaMenuRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.domain;
+package kitchenpos.menus.infrastructure;
 
+import kitchenpos.menus.domain.model.Menu;
+import kitchenpos.menus.domain.model.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
