+++ src/main/java/kitchenpos/shared/external/DefaultPurgomalumClient.java
@@ -1,5 +1,6 @@
-package kitchenpos.infra;
+package kitchenpos.shared.external;
 
+import kitchenpos.shared.external.PurgomalumClient;
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.client.RestTemplate;
