+++ src/main/java/kitchenpos/shared/external/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.infra;
+package kitchenpos.shared.external;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
