+++ src/main/java/kitchenpos/ui/MenuGroupRestController.java
@@ -1,7 +1,7 @@
 package kitchenpos.ui;
 
-import kitchenpos.application.MenuGroupService;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.menus.application.MenuGroupService;
+import kitchenpos.menus.domain.model.MenuGroup;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
