+++ src/test/java/kitchenpos/application/OrderServiceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.application;
 
-import kitchenpos.domain.MenuRepository;
+import kitchenpos.menus.domain.model.MenuRepository;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderRepository;
